Adapt collision tags to jolt collision filters
Tags work exactly as with ODE backend.
Jolt has broad-phase layers and filtering, and object layers for
narrow-phase filtering. Any lovr tag could be either kinematic or
dynamic.

To make this work, each tag is mapped to two object layers, the
kinematic layer and the static layer. Object layers are uint16, so their
value is simply tag << 1, with lowest bit storing the dynamic flag. This
enables easy mapping to broad-phase layer, and much of code around tags
can be reused.

Because of tag-based arithmetics the untagged value cannot be ~0 and is
moved to MAX_TAGS value.